backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=23.0,<24.0
django-cors-headers>=4.3,<5.0
Pillow>=10.0,<11.0
psycopg2-binary>=2.9,<3.0
openai>=1.0,<2.0
gunicorn>=21.2,<22.0
